+++ deltafi-core/src/test/java/org/deltafi/core/converters/EgressFlowPlanConverterTest.java
@@ -59,6 +59,8 @@ class EgressFlowPlanConverterTest {
         assertThat(egressFlow.getValidateActions()).hasSize(1).contains(expectedValidateAction());
         assertThat(egressFlow.getIncludeIngressFlows()).isNull();
         assertThat(egressFlow.getExcludeIngressFlows()).hasSize(3).contains("a", "b", "c");
+        assertThat(egressFlow.getFlowStatus().getState()).isEqualTo(FlowState.STOPPED);
+        assertThat(egressFlow.getFlowStatus().getTestMode()).isFalse();
     }
 
     @Test